airflow_ml_dags/requirements.txt
loguru==0.6.0
tqdm
contourpy==1.0.5
cycler==0.11.0
fonttools==4.37.4
joblib==1.2.0
kiwisolver==1.4.4
matplotlib
numpy
packaging==21.3
pandas
Pillow==9.2.0
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.5
scikit-learn
scipy
seaborn==0.12.1
six==1.16.0
threadpoolctl==3.1.0

